requirements.txt
python-dotenv
Flask
Flask-Bootstrap
Jinja2
Mako
Werkzeug